dummy-with-airflow/requirements.txt
--index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://pypi.org/simple

requests>=2.31
pandas>=2.2
numpy>=1.26
Pillow>=10.3
imagehash>=4.3
minio>=7.2
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
opencv-python-headless>=4.10.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.1
pyarrow>=16.0.0
psycopg2-binary>=2.9
